Update index.md
I've never suggested changes in *anything* on Github (as in, "ever"); but the need for editing here (at lease for the non-technical English-reader) made me intervene...while not really knowing what I'm doing. I only revised the first section of the document, making changes in grammar, syntax, spelling and readability. If the changes are deemed to be unneeded and/or rejected, that's fine; I won't continue. (I also may not continue if the proposed changes *are* accepted, lol.)